[SPARK-56546][SQL] Extract test helpers and minor cleanup
