+++ sql/core/src/test/scala/org/apache/spark/sql/execution/window/SegmentTreeWindowFunctionSuite.scala
@@ -145,7 +145,7 @@ class SegmentTreeWindowFunctionSuite extends QueryTest with SharedSparkSession {
 
     // 2. Directly exercise the frame to confirm the fallback flag flips.
     withSQLConf(enabledConf.toSeq: _*) {
-      SegmentTreeWindowFunctionSuiteHelper.withSmallPartitionFrame(
+      SegmentTreeWindowTestHelpers.withSmallPartitionFrame(
         SQLConf.get, rows = 5) { frame =>
         assert(frame.fallbackUsed,
           "expected fallbackUsed=true for partition smaller than minPartitionRows")
@@ -190,89 +190,5 @@ class SegmentTreeWindowFunctionSuite extends QueryTest with SharedSparkSession {
         .collect().sortBy(_.toString)
     }
     assert(withSegTree.toSeq === baseline.toSeq)
-  }
-}
-
-/**
- * Helper living in the same package so tests can construct a frame directly
- * without going through Catalyst. This mirrors the `private[window]` hook
- * described in contract Section 1.3.
- */
-private object SegmentTreeWindowFunctionSuiteHelper {
-  import org.apache.spark.{SparkEnv, TaskContext}
-  import org.apache.spark.memory.MemoryTestingUtils
-  import org.apache.spark.sql.catalyst.InternalRow
-  import org.apache.spark.sql.catalyst.expressions._
-  import org.apache.spark.sql.catalyst.expressions.aggregate.Sum
-  import org.apache.spark.sql.catalyst.expressions.codegen.GenerateMutableProjection
-  import org.apache.spark.sql.execution.ExternalAppendOnlyUnsafeRowArray
-  import org.apache.spark.sql.types.IntegerType
-
-  def withSmallPartitionFrame(conf: SQLConf, rows: Int)
-      (body: SegmentTreeWindowFunctionFrame => Unit): Unit = {
-    val existing = TaskContext.get()
-    val owned = existing == null
-    val tc = if (owned) {
-      val fake = MemoryTestingUtils.fakeTaskContext(SparkEnv.get)
-      TaskContext.setTaskContext(fake)
-      fake
-    } else existing
-    try {
-      val frame = buildSmallPartitionFrame(conf, rows, tc)
-      try body(frame) finally frame.close()
-    } finally {
-      if (owned) TaskContext.unset()
-    }
-  }
-
-  private def buildSmallPartitionFrame(
-      conf: SQLConf, rows: Int, tc: TaskContext): SegmentTreeWindowFunctionFrame = {
-    val attr = AttributeReference("v", IntegerType, nullable = false)()
-    val input = Seq[Attribute](attr)
-    val fn = Sum(attr)
-    val bufAttrs = fn.aggBufferAttributes
-    val processor = AggregateProcessor(
-      Array[Expression](fn),
-      0,
-      input,
-      (es, s) => GenerateMutableProjection.generate(es, s),
-      Array[Option[Expression]](None))
-    val target = new SpecificInternalRow(Seq(bufAttrs.head.dataType))
-    val array = new ExternalAppendOnlyUnsafeRowArray(
-      tc.taskMemoryManager(),
-      SparkEnv.get.blockManager,
-      SparkEnv.get.serializerManager,
-      tc,
-      1024,
-      SparkEnv.get.memoryManager.pageSizeBytes,
-      Int.MaxValue,
-      Long.MaxValue,
-      Int.MaxValue,
-      Long.MaxValue)
-    val unsafeProj = UnsafeProjection.create(Array(attr.dataType))
-    var i = 0
-    while (i < rows) {
-      val r = new GenericInternalRow(Array[Any](i))
-      array.add(unsafeProj(r))
-      i += 1
-    }
-    val frame = new SegmentTreeWindowFunctionFrame(
-      target,
-      processor,
-      Array(fn),
-      input,
-      RowFrame,
-      RowBoundOrdering(-1),
-      RowBoundOrdering(1),
-      (es, s) => GenerateMutableProjection.generate(es, s),
-      conf,
-      None)
-    frame.prepare(array)
-    var j = 0
-    while (j < rows) {
-      frame.write(j, InternalRow.empty)
-      j += 1
-    }
-    frame
   }
 }
